refactor(slash): move the command handlers next to the registry
The registry lives in `features::slash`; the bodies it dispatches to were still
in main.rs, 1,638 lines of them, which is why "add a slash command" meant editing
a file nobody wants to open. They move to `features::slash_handlers`, so the two
halves of a command — its identity and what it does — are now adjacent.

The cut is one-directional: the module needs 29 helpers from the REPL (which
become `pub(crate)`) and exports 4 back. `fmt_time_ago` deliberately stayed in
main.rs — `core::session_store` uses it, and core must not depend on features.
